requirements.txt
python-dotenv>=1.0.0
anthropic>=0.96.0
requests>=2.33.0
openai>=2.0.0
